requirements.txt
aiohttp==3.9.0b0
discord.py
openai

langchain==0.0.133
python-dotenv==1.0.0
html2text==2020.1.16
numpy==1.24.2
tiktoken==0.3.3
requests==2.28.2
google-api-python-client==2.84.0
openai==0.27.4